fix: created standard component for Switch
