+++ src/components/blocks/LeftNav.js
@@ -1,10 +1,10 @@
 import React, { useEffect, useState } from 'react';
-import { useDispatch } from 'react-redux';
 import styled, { withTheme } from 'styled-components';
 import { ClimateWarehouseLogo } from '../../components';
 import { Link } from 'react-router-dom';
-import { Switch } from '@mui/material';
-import { toggleMode } from '../../store/actions/app';
+import ToggleSwitch from './ToggleSwitch';
+
+
 
 const Container = styled('div')`
   display: flex;
@@ -35,20 +35,10 @@ const MenuItem = styled(Link)`
   text-decoration: none;
 `;
 
-const SwitchContainer = styled('div')`
-  background: ${props => (props.selected ? '#e0f4fe' : 'transparent')};
-  width: 100%;
-  color: ${props => (props.selected ? '#003a8c' : '#ffffff')};
-  font-family: ${props => props.theme.typography.primary.bold};
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  text-decoration: none;
-`;
+
 
 const LeftNav = withTheme(({ children }) => {
   const [location, setLocation] = useState(false);
-  const dispatch = useDispatch();
   useEffect(() => {
     setLocation(window.location.pathname.split(/_(.+)/)[1]);
   }, [window.location]);
@@ -59,14 +49,7 @@ const LeftNav = withTheme(({ children }) => {
         <LogoContainer>
           <ClimateWarehouseLogo />
         </LogoContainer>
-        <SwitchContainer
-          onChange={() => {
-            dispatch(toggleMode);
-          }}>
-          <div>Warehouse</div>
-          <Switch />
-          <div>Registry</div>
-        </SwitchContainer>
+        <ToggleSwitch />
         <MenuItem
           selected={location === 'projects'}
           to="/projects"
